installer/gcp/scripts/python/requirements.txt
google-cloud-storage
google-api-python-client
oauth2client
termcolor
